cambios

+++ facturacion/src/main/java/com/facturacion/facturacion/Repository/CarritoRepository.java
@@ -9,7 +9,7 @@ import com.facturacion.facturacion.Model.CarritoModel;
 import com.facturacion.facturacion.Model.ClientsModel;
 
 @Repository
-public interface CarritoRepository extends CrudRepository<CarritoModel,Long>{
+public interface CarritoRepository extends CrudRepository<CarritoModel,Integer>{
 
     public abstract Optional<CarritoModel> findByClientsModel(ClientsModel clientsModel);
